requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
anyascii==0.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
exceptiongroup==1.2.2
fastapi==0.104.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jira==3.8.0
kiwisolver==1.4.8
langdetect==1.0.9
MarkupSafe==3.0.2
matplotlib==3.10.0
mplcursors==0.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==24.3.1
psutil==6.1.1
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-doctr==0.11.0
python-multipart==0.0.6
pytz==2025.1
PyYAML==6.0.2
qwen-vl-utils==0.0.10
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
safetensors==0.5.2
scipy==1.15.2
seaborn==0.13.2
setuptools==75.6.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
sympy==1.13.1
together==1.4.6
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers>=4.37.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.86
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.24.0
wheel==0.45.1
python-dotenv==1.0.0